add rss-service.ts and clean code
