+++ src/views/404.ejs
@@ -1,0 +1,11 @@
+<!DOCTYPE html>
+<html lang="fr">
+<head>
+    <meta charset="utf-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1">
+    <title>Node.js</title>
+</head>
+<body>
+    <h1>oiefjsdjfoisdjfiosdjfosdjfsdf</h1>
+</body>
+</html>
